test(analyzer): 收敛 Core.Tests 与 Cqrs.Tests 警告批次
- 拆分 GameContextTests、ArchitectureServicesTests、RegistryInitializationHookBaseTests 与 Cqrs 测试辅助类型，消除批次内 MA0048 热点
- 修复 Core.Tests 零散可空性、集合抽象和测试辅助 warning，使受影响 Release 构建清零
- 更新 analyzer-warning-reduction 跟踪与 trace，记录 236 条仓库根 warning 基线和 45/50 停止点
