+++ GFramework.Core.Tests/Architectures/AnotherTestArchitectureContext.cs
@@ -1,0 +1,8 @@
+namespace GFramework.Core.Tests.Architectures;
+
+/// <summary>
+///     表示用于验证上下文类型不匹配分支的测试架构上下文。
+/// </summary>
+public class AnotherTestArchitectureContext : TestArchitectureContext
+{
+}
