+++ GFramework.Core.Tests/Events/EmptyEvent.cs
@@ -1,0 +1,6 @@
+namespace GFramework.Core.Tests.Events;
+
+/// <summary>
+///     表示不携带任何负载的空测试事件。
+/// </summary>
+public sealed class EmptyEvent;
