+++ GFramework.Core.Tests/Logging/LoggerTests.cs
@@ -419,44 +419,3 @@ public class LoggerTests
         Assert.That(fatalLogger.Logs[0].Level, Is.EqualTo(LogLevel.Fatal));
     }
 }
-
-/// <summary>
-///     测试用的日志记录器实现类，继承自AbstractLogger
-/// </summary>
-public sealed class TestLogger : AbstractLogger
-{
-    private readonly List<LogEntry> _logs = new();
-
-    /// <summary>
-    ///     初始化TestLogger的新实例
-    /// </summary>
-    /// <param name="name">日志记录器的名称，默认为null</param>
-    /// <param name="minLevel">最小日志级别，默认为LogLevel.Info</param>
-    public TestLogger(string? name = null, LogLevel minLevel = LogLevel.Info) : base(name, minLevel)
-    {
-    }
-
-    /// <summary>
-    ///     获取按写入顺序保存的日志条目只读视图
-    /// </summary>
-    public IReadOnlyList<LogEntry> Logs => _logs;
-
-    /// <summary>
-    ///     将日志信息写入内部存储
-    /// </summary>
-    /// <param name="level">日志级别</param>
-    /// <param name="message">日志消息</param>
-    /// <param name="exception">相关异常（可选）</param>
-    protected override void Write(LogLevel level, string message, Exception? exception)
-    {
-        _logs.Add(new LogEntry(level, message, exception));
-    }
-
-    /// <summary>
-    ///     表示单个日志条目的记录类型
-    /// </summary>
-    /// <param name="Level">日志级别</param>
-    /// <param name="Message">日志消息</param>
-    /// <param name="Exception">相关异常（可选）</param>
-    public sealed record LogEntry(LogLevel Level, string Message, Exception? Exception);
-}
