+++ GFramework.Core.Tests/Resource/TestResource.cs
@@ -1,0 +1,17 @@
+namespace GFramework.Core.Tests.Resource;
+
+/// <summary>
+///     表示 ResourceManager 测试使用的简单资源对象。
+/// </summary>
+public class TestResource
+{
+    /// <summary>
+    ///     获取或设置资源内容。
+    /// </summary>
+    public string Content { get; set; } = string.Empty;
+
+    /// <summary>
+    ///     获取或设置一个值，指示资源是否已经被测试加载器标记为已卸载。
+    /// </summary>
+    public bool IsDisposed { get; set; }
+}
